// src/nft/dto/get-collections.dto.ts
export class GetCollectionsResponse {
  _id: string;
  symbol: string;
  chain: string;
  type: string;
  transactionhash: string;
  baseuri: string;
  imageuri: [string];
  __v: 0;
  description: string;
  collection_name: string;
  contract_address: string;
  owner_address: string;
  updatedAt: string;
}